refactor: now fabric nodes and controllers can use names instead of ids, common code in module_utils/vmware.py
