+++ module_utils/vmware.py
@@ -14,7 +14,60 @@ import json
 from ansible.module_utils.urls import open_url, fetch_url
 from ansible.module_utils.six.moves.urllib.error import HTTPError
 from ansible.module_utils._text import to_native
+import requests, ssl
+from pyVim.connect import SmartConnect
 
+def find_morefId(obj_name, obj_list):
+    """
+    Gets an object out of a list (obj_list) whos name matches obj_name.
+    """
+    for o in obj_list:
+        if o.name == obj_name:
+            return o._moId
+    raise Exception("Unable to find object ", obj_name)
+
+def find_moref_ids_for_deployment(vm_deployment_config, vc_host, vc_username, vc_password, vc_datacenter):
+    requests.packages.urllib3.disable_warnings()
+    ssl._create_default_https_context = ssl._create_unverified_context
+    context = ssl.create_default_context()
+    context.check_hostname = False
+    context.verify_mode = ssl.CERT_NONE
+
+    deploy_config = dict()
+
+    si = SmartConnect(host=vc_host, user=vc_username, pwd=vc_password)
+    content = si.RetrieveContent()
+    datacenter_obj = None
+
+    # select the right datacenter using its name
+    datacenter_list = content.rootFolder.childEntity
+    for datacenter in datacenter_list:
+        if datacenter.name == vc_datacenter:
+            datacenter_obj = datacenter
+        else:
+            raise Exception("Datacenter not found", vm_deployment_config['compute_id'])
+
+    # translate datastore name
+    datastore_list = datacenter_obj.datastoreFolder.childEntity
+    deploy_config['storage_id'] = find_morefId(vm_deployment_config['storage_id'], datastore_list)
+
+    # translate cluster name (compute_id)
+    cluster_list = datacenter_obj.hostFolder.childEntity
+    deploy_config['compute_id'] = find_morefId(vm_deployment_config['compute_id'], cluster_list)
+
+    # translate all data networks
+    deploy_config['data_network_ids'] = []
+    network_list = datacenter_obj.networkFolder.childEntity
+
+    if 'data_network_ids' in vm_deployment_config: 
+        for network in vm_deployment_config['data_network_ids']:
+            deploy_config['data_network_ids'].append(find_morefId(network, network_list))
+            #data_network_ids.append(find_morefId(network, network_list))
+
+    # translate management network
+    deploy_config['management_network_id'] = find_morefId(vm_deployment_config['management_network_id'], network_list)
+
+    return deploy_config
 
 def vmware_argument_spec():
     return dict(
@@ -66,6 +119,7 @@ def get_params(args={}, args_to_remove=[]):
         if value is None:
             args.pop(key, None)
     return args
+
 
 
 def api_call(module, api_path, method='GET', data=None, headers={}):
